agents/crewai/requirements.txt
crewai
crewai==0.152.0          
pyyaml>=6.0
openai>=1,<2
gitpython>=3.1
pygithub>=2.3
pytest
langchain==0.2.0     